Mass Shootings
This repo shows a set of SAS advanced steps which include PROC SQL and SAS MACRO's  that I used to tackle the  "Mass Shootings" dataset. This analysis provides comprehensive information regarding Mass Shootings from 1982-2017 i.e," frequency of occurrence, weapons used, possession of weapons (legal/illegal) and mental health condition of the shooter along with the information regarding the number of people dead and injured in every incident

MACRO's DESCRIPTION:
To start with Macros variables were created to save the path of the files imported as text and then the text is substituted in the libname and filename statements to assign library.
Automatic Macro variables &syslibrc, &syserr, &sysparm with the %put statement were used to check if the library is successfully assigned, they write value 0 to the SAS log if the library is successfully assigned and a value of 1 if the library assignment is failed. User-defined variables were created and saved to automate the PROC SQL steps. Macro definitions were created using %MACRO and %MEND statements for the SAS code that is repeated. For Example: creating a table that shows a total number of victims each year. System options like Mcompilenote = all is used, which writes the compilation information to the SAS log. To continue with, the macro definition is called is called along with positional parameters to create reports for each year from 1982 to 2017. Mprint system option is used along with the call, which displays the code that is sent for compiling in the log and the sas statements displayed by macro execution.
To decrease the number of macros functions or to limit the number of the macro functions %if %then statement and %do loop is used to write complete and partial SAS steps on conditions along with If then clauses. MLOGIC system option is used while calling a macro function which writes to SAS log the information about macros resolved during execution. Mlogic option identifies the macro execution path, which helps us to debug the problem.

SQL DESCRIPTION:
 	The Process of analysis starts with combining datasets using INNER JOIN, specifying the variables that are needed for the analysis from the two tables. PROC SQL step with summary function count has been used on dataset MG_VICTIMS, to count the number of victims per every year from 1982 to 2017. To continue with PROC SQL step with summary function count is used on dataset MG_YEAR, MG_STATE, to count the no of incidents per each year, to gives us a report of frequency of occurrence of incidents every year, to produces a report which has the information regarding number many incidents occurred in each state, to create columns that count number of weapons obtained legally and illegally, to create column that has count weather mass shooter has any mental conditions.  
New columns were created using a summary function to produce the count of weapons obtained legally and illegally, count of male and female shooters and the information regarding whether the mass shooter has any previous mental condition. CREATE TABLE is used to create a new table called weapons which stores the count of different weapons used and no of weapons used in each incident by the mass shooter. A case expression is used to designate levels according to the weapons used by mass shooters for each incident.


CONCLUSION:
	The analysis using PROC SQL and Macros has helped the Shiksha organization to draw the following conclusions. Analyzing a total number of incidents every year has shown that during 1980 to 1990 the frequency of mass shooting were two shootings per year, for 2017 they have increased to 1 incident per month. The analysis of MG_Gender has shown 99% of mass shooters have been Males. Analyzing MG_weapons data have shown that most of the weapons used by the mass shooters are legally obtained. Most of the weapons used for semiautomatic rifles and
Assault_riffiles.   In 60% of the incidents, mass shooters have used more than two weapons, which shows that the acts of mass shootings were planned and are not spontaneously acted.
Analysis the MG_Victims data, the conclusion was drawn was 60% of the mass shooters have pre-mental conditions. Out of 150 weapons used in 91 incidents 48 weapons were banned weapons. Even though these mental conditions don’t proof the mass shooters are mentally disabled it shows all these shooters have domestic violence records. Analyzing the MG_States shows incidents have occurred frequently in California, Florida and Ohio states. Schools, business areas, and public places are prone to more mass shootings in last three decades. We are at more risk of mass shootings in these places.